Refactored IncomingMessages to clean architecture
