+++ source/IncomingMessages.Domain/MessageParsers/RSM016/AggregatedMeasureDataXmlMessageParser.cs
@@ -14,7 +14,7 @@
 
 using System.Xml.Linq;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Abstractions;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
 using Energinet.DataHub.EDI.IncomingMessages.Domain.Schemas.Cim.Xml;
 
 namespace Energinet.DataHub.EDI.IncomingMessages.Domain.MessageParsers.RSM016;
