+++ source/IncomingMessages.Domain/Models/IncomingMessage/IDelegatedIncomingMessageSeries.cs
@@ -11,10 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 
-namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Abstractions;
+namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
 
 /// <summary>
 /// Represents a delegated incoming message series, with the data needed for a delegation
