+++ source/IncomingMessages.Domain/Models/IncomingMessage/IIncomingMessage.cs
@@ -12,10 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.BuildingBlocks.Domain.DataHub;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 
-namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Abstractions;
+namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
 
 /// <summary>
 /// Represents an incoming message
