+++ source/IncomingMessages.Domain/Models/IncomingMessage/IIncomingMessageSeries.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
 
-namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Abstractions;
+namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
 
 /// <summary>
 /// Represents a series of an incoming message
