+++ source/IncomingMessages.Domain/Models/MessageId/IMessageIdRepository.cs
@@ -12,19 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
-
-namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Validation;
+namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Models.MessageId;
 
 /// <summary>
-/// Validation for Message Type
+/// Store containing message id for all received market documents
 /// </summary>
-public interface IMessageTypeValidator
+public interface IMessageIdRepository
 {
     /// <summary>
-    /// Validates Message Type
+    /// Store message id for the sender
     /// </summary>
-    /// <param name="message"></param>
-    /// <param name="cancellationToken"></param>
-    public Task<Result> ValidateAsync(IIncomingMessage message, CancellationToken cancellationToken);
+    Task AddAsync(string senderNumber, string messageId, CancellationToken cancellationToken);
+
+    /// <summary>
+    /// Checks if <paramref name="messageId"/> is already registered by the sender <paramref name="senderNumber"/>
+    /// </summary>
+    Task<bool> MessageIdExistsAsync(string senderNumber, string messageId, CancellationToken cancellationToken);
 }
