+++ source/IncomingMessages.Domain/Validation/IProcessTypeValidator.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Abstractions;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
 
 namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Validation;
