+++ source/IncomingMessages.Domain/Validation/ISenderAuthorizer.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.EDI.IncomingMessages.Domain.Abstractions;
+using Energinet.DataHub.EDI.IncomingMessages.Domain.Models.IncomingMessage;
 
 namespace Energinet.DataHub.EDI.IncomingMessages.Domain.Validation;
