+++ source/IncomingMessages.Infrastructure/IncomingMessages.Infrastructure.csproj
@@ -13,6 +13,8 @@
   <ItemGroup>
     <ProjectReference Include="..\BuildingBlocks.Domain\BuildingBlocks.Domain.csproj" />
     <ProjectReference Include="..\BuildingBlocks.Infrastructure\BuildingBlocks.Infrastructure.csproj" />
+    <ProjectReference Include="..\DataAccess\DataAccess.csproj" />
+    <ProjectReference Include="..\IncomingMessages.Application\IncomingMessages.Application.csproj" />
     <ProjectReference Include="..\IncomingMessages.Domain\IncomingMessages.Domain.csproj" />
     <ProjectReference Include="..\IncomingMessages.Interfaces\IncomingMessages.Interfaces.csproj" />
     <ProjectReference Include="..\Process.Interfaces\Process.Interfaces.csproj" />
